Fix #115 mmap length greater than file size
Previously there was a race condition where after `os.ftruncate()`
the size of the new file wouldn't be updated by the time `mmap.mmap()`
was called. An fsync seems to be the solution[1].

Also resolved an issue where open files were not closed. Not all of the
file leaks have been resolved, there is a TODO describing the problem better.

[1] https://lwn.net/Articles/457667/ see "When Should You Fsync?"