Refactor HolodeckEnvironments, Holodeck.make()
These classes had an inconsistent division of responsibilities. Previously both would call functions in the packagemanager to search for config files, and there where several ways to specify which scenario to load.

Now, `HolodeckEnvironment` does no parsing of configs, it expects to be provided a dictionary for the scenario config. This makes it so you can define your scenario as a python dictionary instead of a JSON file located in the holodeck folder.