Fix the max open files error
It turns out that calling os.remove or unlinking a
mapped memory file might delete the actual file
but the OS keeps the mapped file open until there's
no reference to that memory block. We were removing
the mapped memory files we created, but since we
still kept references to the location in memory,
the OS did not close the file (handle). Running multiple
Holodeck instances eventually in one process eventually
exaust the open file limit.

This closes #356 and also closes #407